requirements.txt
python-dotenv
opencv-python
ultralytics
twilio
requests
python-telegram-bot
